+++ src/barchart/qpercentbarseries.cpp
@@ -89,8 +89,8 @@ void QPercentBarSeriesPrivate::scaleDomain(Domain& domain)
     int tickYCount(domain.tickYCount());
 
     qreal x = categoryCount();
-    minX = qMin(minX, -0.5);
-    maxX = qMax(maxX, x - 0.5);
+    minX = qMin(minX, - (qreal)0.5);
+    maxX = qMax(maxX, x - (qreal)0.5);
     minY = 0;
     maxY = 100;
     tickXCount = x+1;
