+++ src/barchart/qstackedbarseries.cpp
@@ -91,9 +91,9 @@ void QStackedBarSeriesPrivate::scaleDomain(Domain& domain)
 
     qreal x = categoryCount();
     qreal y = maxCategorySum();
-    minX = qMin(minX, -0.5);
+    minX = qMin(minX, - (qreal)0.5);
     minY = qMin(minY, y);
-    maxX = qMax(maxX, x - 0.5);
+    maxX = qMax(maxX, x - (qreal)0.5);
     maxY = qMax(maxY, y);
     tickXCount = x+1;
